+++ examples/test/repo-question.test.tsx
@@ -1,9 +1,9 @@
 import { test } from "vitest";
 import { Performer } from "@performer/core";
-import { RecentWork } from "../src/repo-question/index.js";
+import { App } from "../src/repo-question/index.js";
 
 test("should answer question about openai github repos", async () => {
-  const performer = new Performer(<RecentWork user="openai" />);
+  const performer = new Performer(<App user="openai" />);
   performer.logConfig.showDeltaEvents = false;
   performer.start();
   await performer.waitUntilSettled();