+++ packages/performer-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @performer/core
 
+## 0.0.17
+
+## 0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- be2f237: Add getCurrentMessages() helper to Performer class
+
+## 0.0.14
+
 ## 0.0.13
 
 ### Patch Changes